chore: Avoid circular dependency
The *Snapping.java and *Route.java class had a circular import pattern with their respective CLI classes. This is now broken up by excluding the main function code.
